pyproject.toml
[project]
name = "openai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.27",
    "openai-agents>=0.1.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]
